requirements.txt
Flask~=3.0.3
Flask-Login~=0.6.3
bcrypt~=4.2.0
SQLAlchemy~=2.0.35
python-dotenv~=1.0.1
PyYAML~=6.0.2
Werkzeug~=3.0.4
Flask-Migrate~=4.0.7